// frontend/src/app/api/generated/model/additionalRuleGroupDto.ts
/**
 * RecurringTransactionsAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RuleDto } from './ruleDto';


export interface AdditionalRuleGroupDto { 
    id: string;
    name: string;
    normalizedName: string;
    rules: Array<RuleDto>;
    excludedTransactionCount: number;
    createdAt: string;
    updatedAt: string;
}