// frontend/src/app/api/generated/model/createRuleRequest.ts
/**
 * RecurringTransactionsAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RuleType } from './ruleType';
import { TargetField } from './targetField';


export interface CreateRuleRequest { 
    ruleType: RuleType;
    targetField?: TargetField;
    text?: string;
    strict?: boolean;
    threshold?: number;
    amount?: number;
    fluctuationRange?: number;
}