// frontend/src/app/api/generated/model/simulationDraftType.ts
/**
 * RecurringTransactionsAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export enum SimulationDraftType {

    RecurringPayment = 'RECURRING_PAYMENT',

    AdditionalGroup = 'ADDITIONAL_GROUP'
}