// frontend/src/app/api/generated/model/recalculationSummaryResponse.ts
/**
 * RecurringTransactionsAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RecalculationSummaryResponse { 
    transactionsMarkedInterAccount: number;
    transactionLinksRemoved: number;
    recurringPaymentsDeleted: number;
    recurringPaymentsDetected: number;
}